\name{m}
\alias{m}
double2
\title{
multiply
}
\description{
test run double
}
\usage{
m(a,...)
}
value to double
\arguments{ na2
  \item{a}{
numeric
}
}
\details{
NA
}
\value{
test
}
\references{
shen
}
\author{
sheen
}
\note{
ho
}


\seealso{
gffgg
}
\examples{

d 3;
function (a) 
{
    s <- a + a
    return(s)
  }
}
double
\keyword{test3 }
\keyword{ f-test }
